refactor planner to add query preprocessor functionality
